implement #111: Add groups to parser
`group` can now be used to declare a group of actions.  This can be used
to group together a specific set of actions.  But also to make
binding/relation groups.  There is the convenient `ungroup GROUP` that
removes all bindings/associations a group shallowly created.  For each
relation is simply does an unrelate with the same pattern.  And for each
bind a simple unbind.

+++ include/core/action.h
@@ -7,6 +7,8 @@
  * The user can invoke any actions in any order at any time.
  */
 
+#include <stddef.h>
+
 #include "bits/actions.h"
 
 /* action codes */
@@ -32,7 +34,7 @@ struct action_list {
         unsigned data_count;
     } *items;
     /* the number of items in `items` */
-    unsigned number_of_items;
+    size_t number_of_items;
     /* the data associated to the actions */
     struct parse_data *data;
 };
